pyproject.toml
[project]
name = "company_structure_api"
version = "0.1.0"
description = "A Python package to download UK Companies House data into a searchable DuckDB database with FTS."
requires-python = ">=3.11"
authors = [
    { name = "Your Name", email = "your@email.com" },
]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Topic :: Database",
    "Topic :: Utilities",
]
dependencies = [
    "duckdb>=0.10.0",
    "rich>=13.0.0", # For nice progress bars during download
    "duckdb-extension-fts",
    "pydantic>=2.10.6",
    "httpx>=0.28.1",
    "fastapi>=0.116.2",
    "jinja2>=3.1.6",
    "pyarrow>=17.0.0",
    "openai>=1.2.3",
    "pydantic-settings>=2.10.1",
    "pydantic-settings>=2.8.1",
    "uvicorn>=0.35.0",
]

[project.scripts]
company_structure_api = "company_structure_api:main"

[project.urls]
Homepage = "https://github.com/your-username/companies-house-db"

[tool.pytest.ini_options]
# Register the 'manual' marker to avoid warnings and add a description.
pythonpath = [
  "src"
]
markers = [
    "manual: marks a test as manual to run on demand (skipped by default)",
]
# Add a default option to exclude tests marked as 'manual'.
addopts = "-m 'not manual'"

[dependency-groups]
dev = [
    "trio>=0.27.0",
    "pytest-trio>=0.8.0",
    "requests>=2.28.0",
    "pytest-mock>=3.14.0",
    "rsconnect-python>=1.27.1",
]

[build-system]
requires = ["uv_build>=0.8.17,<0.9.0"]
build-backend = "uv_build"
